Set chrony_version to nil if not installed
Previously when chrony_version is passed to Gem::Version.new in
chronyd.conf.erb, it was possible to cause an error if chrony wasn't yet
installed (a fresh deployment). This tests if chrony exists before
letting rpm set the fact to 'package chrony is not installed' (not a
valid version string).